+++ src/txBuilders/loyaltyProgram/index.ts
@@ -1,6 +1,7 @@
 import { Transaction } from '@mysten/sui/transactions';
 import { SuiTxBlock as SuiKitTxBlock } from '@scallop-io/sui-kit';
 import { ScallopBuilder } from 'src/models/index.js';
+import type { ReadTransport } from 'src/models/index.js';
 import {
   LoyaltyProgramTxBlock,
   ScallopTxBlock,
@@ -21,7 +22,7 @@ import type { MoveCallContext } from '../context.js';
  * @return Scallop loyalty program txBlock.
  */
 export const newLoyaltyProgramTxBlock = (
-  builder: ScallopBuilder,
+  builder: ScallopBuilder<ReadTransport>,
   initTxBlock?: ScallopTxBlock | SuiKitTxBlock | Transaction
 ) => {
   const txBlock =